qwen_agent_poc/requirements.txt
qwen-agent>=0.0.9
gradio>=4.0.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.0.0 